+++ src/main/java/com/hoangluongtran0309/releaseflow/release/ReleaseNotFoundException.java
@@ -1,0 +1,8 @@
+package com.hoangluongtran0309.releaseflow.release;
+
+public class ReleaseNotFoundException extends RuntimeException {
+
+    public ReleaseNotFoundException() {
+        super("Release was not found.");
+    }
+}