+++ src/main/java/com/hoangluongtran0309/releaseflow/release/ReleaseStatus.java
@@ -1,0 +1,5 @@
+package com.hoangluongtran0309.releaseflow.release;
+
+public enum ReleaseStatus {
+    DRAFT
+}